+++ src/hooks/index.ts
@@ -1,8 +1,9 @@
 export { useActiveUser } from './use-active-user';
+export { useLogin } from './use-login';
 export { useNdk } from './use-ndk';
 export { useNewEvent } from './use-new-event';
-export { useNip07 } from './use-nip07';
 export { useNostrHooks } from './use-nostr-hooks';
 export { useProfiles } from './use-profiles';
 export { usePublish } from './use-publish';
+export { useSigner } from './use-signer';
 export { useSubscribe } from './use-subscribe';