+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Nostr-Hooks
 
+## 2.7.0
+
+- Improved interacting with different signers.
+- Added `useSigner` hook.
+- Added `useLogin` hook.
+- Removed `useNip07` hook.
+
+### Breaking Changes
+
+- Extracted signer related functionalities from `useNDK` hook to `useSigner` hook. Now you need to import `setSigner` method from `useSigner` hook instead of `useNDK` hook.
+- Removed `useNip07` hook. Now you can use `useLogin` hook to login with extension (NIP-07).
+
 ## 2.6.0
 
 - Added `fetchProfiles` boolean to `useSubscribe` hook to fetch profiles on subscription.